requirements.txt
opencv-python>=4.10.0
numpy>=1.26.0
Pillow>=10.0.0
streamlit>=1.38.0
scikit-image>=0.24.0
reportlab>=4.2.0
pandas>=2.2.0
matplotlib>=3.8.0
